refactor: ♻️ split agent_session_service into query/command submodules (#263)
